+++ tests/projects/rust/rust_call_cxx_library/src/foo.cc
@@ -1,4 +1,3 @@
-extern "C" int add(int a, int b)
-{
+extern "C" int add(int a, int b) {
     return a + b;
 }