+++ xmake/rules/rust/xmake.lua
@@ -32,10 +32,12 @@ rule("rust.cxxbridge")
 rule("rust.build")
     set_sourcekinds("rc")
     on_load(function (target)
+        -- set cratetype
         local cratetype = target:values("rust.cratetype")
         if cratetype == "staticlib" then
             assert(target:is_static(), "target(%s) must be static kind for cratetype(staticlib)!", target:name())
             target:add("arflags", "--crate-type=staticlib")
+            target:data_set("inherit.links.exportlinks", false)
         elseif cratetype == "cdylib" then
             assert(target:is_shared(), "target(%s) must be shared kind for cratetype(cdylib)!", target:name())
             target:add("shflags", "--crate-type=cdylib")
@@ -55,6 +57,10 @@ rule("rust.build")
         elseif target:is_binary() then
             target:add("ldflags", "--crate-type=bin")
         end
+
+        -- set edition
+        local edition = target:values("rust.edition") or "2018"
+        target:add("rcflags", "--edition", edition, {force = true})
     end)
     on_build("build.target")