+++ xmake/rules/utils/inherit_links/inherit_links.lua
@@ -80,11 +80,13 @@ function main(target)
         -- @note we only export links for static target,
         -- and we need pass `{public = true}` to add_packages/add_links/... to export it if want to export links for shared target
         --
-        if targetkind == "static" then
-            for _, name in ipairs({"rpathdirs", "frameworkdirs", "frameworks", "linkdirs", "links", "syslinks"}) do
-                local values = _get_values_from_target(target, name)
-                if values and #values > 0 then
-                    target:add(name, values, {public = true})
+        if target:data("inherit.links.exportlinks") ~= false then
+            if targetkind == "static" then
+                for _, name in ipairs({"rpathdirs", "frameworkdirs", "frameworks", "linkdirs", "links", "syslinks"}) do
+                    local values = _get_values_from_target(target, name)
+                    if values and #values > 0 then
+                        target:add(name, values, {public = true})
+                    end
                 end
             end
         end