+++ xmake/templates/rust/static/project/src/main.rs
@@ -1,7 +1,6 @@
 extern crate foo;
 
-fn main()
-{
+fn main() {
     println!("hello xmake!");
     println!("add: {}", foo::add(1, 1));
 }